--- src/codegen/index.ts
@@ -1,6 +1,0 @@
-import { Http } from './http';
-
-const createHttp = (...args: ConstructorParameters<typeof Http>) =>
-  new Http(...args);
-
-export { createHttp };
