+++ src/lib/object.ts
